+++ sources/common.h
@@ -39,7 +39,7 @@ char *getCurrentTime();
 
 char *strMallocCat(const char *str1, const char *str2);
 
-void freePointer(void **pointer, int *isMalloc);
+void freePointer(void **pointer, short *isMalloc);
 
 //int getCountListMimeType();
 //
